Fever API: move HandleRequest file
